% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cor2cov}
\alias{cor2cov}
\title{Correlation to Covariance}
\usage{
cor2cov(cor, sd)
}
\arguments{
\item{cor}{Numeric matrix.
A \eqn{p \times p} positive definite correlation matrix.}

\item{sd}{Numeric vector.
A vector of \eqn{p} standard deviations.}
}
\value{
Returns a covariance matrix.
}
\description{
Converts a correlation matrix to a covariance matrix.
}
\examples{
R <- matrix(
  data = c(
    1, 0.509902, 0.26,
    0.509902, 1, 0.509902,
    0.26, 0.509902, 1
  ),
  ncol = 3
)
cor2cov(
  cor = R,
  sd = c(15, 15, 15)
)
}
\seealso{
Other matrix functions: 
\code{\link{low2sym}()},
\code{\link{tr}()},
\code{\link{up2sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{matrix functions}
\keyword{matrix}
