% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.diag}
\alias{is.diag}
\title{Diagonal Matrix}
\usage{
is.diag(X, tol = 1e-08, stop = FALSE)
}
\arguments{
\item{X}{Square numeric matrix.}

\item{tol}{Numeric.
Tolerance.}

\item{stop}{Logical.
If \code{TRUE},
stops and returns an error if \code{X} is not a diagonal matrix.
If \code{FALSE},
returns the result of the test.}
}
\value{
If \code{stop = FALSE},
returns \code{TRUE},
if the sum of the off-diagonal elements of \code{X} is equal to zero, or
returns \code{FALSE},
if the sum of the off-diagonal elements of \code{X} is NOT equal to zero.
}
\description{
Checks if off-diagonal elements of a square matrix are all zeroes.
}
\details{
The off-diagonal elements are added and compared to a tolerance value.
If the sum is less than or equal to the tolerance value,
all the elements are assumed to be zeroes.
}
\examples{
A <- matrix(
  data = 1:9,
  nrow = 3
)
B <- matrix(
  data = 1:10,
  ncol = 2
)
C <- diag(3)
# Returns TRUE
is.diag(X = C)
# Returns FALSE
is.diag(X = A)
is.diag(X = B)
}
\references{
\href{https://en.wikipedia.org/wiki/Diagonal_matrix}{Wikipedia: Diagonal matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
