% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.invertible}
\alias{is.invertible}
\title{Invertible Matrix}
\usage{
is.invertible(X, tol = 1e-08, stop = FALSE)
}
\arguments{
\item{X}{Numeric matrix.
Square matrix.}

\item{tol}{Numeric.
Tolerance.}

\item{stop}{Logical.
If \code{TRUE},
stops and returns an error
with any of the following conditions:
\code{X} is not a square matrix, or
\code{X} is not invertible, that is, a singular matrix.}
}
\value{
Returns \code{TRUE}
if the matrix is invertible.
Returns \code{FALSE}
if the matrix is not invertible, that is, singular.
}
\description{
Checks if a square matrix is invertible.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is.invertible(X = Sigma)
}
\references{
\href{https://en.wikipedia.org/wiki/Invertible_matrix}{Wikipedia: Invertible Matrix}
\href{https://en.wikipedia.org/wiki/Singular_matrix}{Wikipedia: Singular Matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{matrix}
