% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.positive.definite}
\alias{is.positive.definite}
\title{Positive Definite}
\usage{
is.positive.definite(X, tol = 1e-08, stop = FALSE)
}
\arguments{
\item{X}{Numeric matrix.
Symmetric matrix.}

\item{tol}{Numeric.
Tolerance.}

\item{stop}{Logical.
If \code{TRUE},
stops and returns an error
with any of the following conditions:
\code{X} is not a square matrix,
\code{X} is not a symmetric matrix, or
\code{X} is not a positive definite matrix.}
}
\value{
Returns \code{TRUE}
if the matrix is positive definite,
that is,
all eigenvalues are positive.
Returns \code{FALSE}
if the matrix is not positive definite,
that is,
any of the eigenvalues are are less than or equal to zero.
}
\description{
Checks if eigenvalues in a square matrix are positive.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is.positive.definite(X = Sigma)
}
\references{
\href{https://en.wikipedia.org/wiki/Definiteness_of_a_matrix}{Wikipedia: Definiteness of a Matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{matrix}
