% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{is.singular.matrix}
\alias{is.singular.matrix}
\title{Singular Matrix}
\usage{
is.singular.matrix(x, tol = 1e-08)
}
\arguments{
\item{x}{A numeric square matrix.}

\item{tol}{Tolerance.}
}
\value{
Returns \code{TRUE}
if the matrix is singular or not invertible,
meaning the determinant of the matrix is 0.
}
\description{
Checks if a square matrix is not invertible.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is.singular.matrix(Sigma)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{matrix}
