% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.square}
\alias{is.square}
\title{Square Matrix}
\usage{
is.square(X, stop = FALSE)
}
\arguments{
\item{X}{Numeric matrix.}

\item{stop}{Logical.
If \code{TRUE},
stops and returns an error if \code{X} is not a square matrix.
If \code{FALSE},
returns the result of the test.}
}
\value{
If \code{stop = FALSE},
returns \code{TRUE},
if the number of row and columns in \code{X} are equal, or
returns \code{FALSE},
if the number of row and columns in \code{X} are not equal.
}
\description{
Checks if a matrix is square \eqn{\left( A_{m \times m} \right)} .
}
\examples{
A <- matrix(
  data = 1:9,
  nrow = 3
)
dim(A)
B <- matrix(
  data = 1:10,
  ncol = 2
)
dim(B)
# Returns TRUE
is.square(X = B)
# Returns FALSE
is.square(X = A)
}
\references{
\href{https://en.wikipedia.org/wiki/Square_matrix}{Wikipedia: Square matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
