% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.symmetric}
\alias{is.symmetric}
\title{Symmetric Matrix}
\usage{
is.symmetric(X, stop = FALSE)
}
\arguments{
\item{X}{Numeric matrix.
A \eqn{p \times p} matrix.}

\item{stop}{Logical.
If \code{TRUE},
stops and returns an error
with any of the following conditions:
\code{X} is not a square matrix or
\code{X} is not a symmetric matrix.}
}
\value{
Returns \code{FALSE},
if the \code{X} is not square and not symmetric.
Returns \code{TRUE},
if the \code{X} is symmetric.
}
\description{
Checks if a matrix is symmetric.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is.symmetric(X = Sigma)
}
\author{
Ivan Jacob Agaloos Pesigan
}
