% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{low2sym}
\alias{low2sym}
\title{Lower Triangle to Symmetric}
\usage{
low2sym(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
Returns a symmetric matrix.
}
\description{
Creates a symmetric matrix from the lower triangle
of a square matrix.
}
\examples{
x <- matrix(NA, ncol = 4, nrow = 4)
x[lower.tri(x, diag = TRUE)] <- 1:10
low2sym(x = x)
}
\seealso{
Other matrix functions: 
\code{\link{cor2cov}()},
\code{\link{tr}()},
\code{\link{up2sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{matrix functions}
\keyword{matrix}
