% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{reindex}
\alias{reindex}
\title{Reindex a Symmetric Matrix.}
\usage{
reindex(X, index)
}
\arguments{
\item{X}{Numeric matrix.
Symmetric matrix.}

\item{index}{Vector of integers.
Index used to reindex the \code{X}.}
}
\value{
Returns a reindexed symmetric matrix.
}
\description{
Reindex a symmetric matrix.
}
\references{
\href{https://en.wikipedia.org/wiki/Permutation_matrix}{Wikipedia: Permutation Matrix}
}
