% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{tr}
\alias{tr}
\title{Trace of a square matrix}
\usage{
tr(x)
}
\arguments{
\item{x}{A \eqn{p \times p} matrix.}
}
\value{
Returns the sum of the diagonal of a square matrix.
}
\description{
Sum of the diagonal of a square matrix.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
tr(Sigma)
}
\references{
\href{https://en.wikipedia.org/wiki/Trace_(linear_algebra)}{Wikipedia: Trace (Linear Algebra)}
}
\seealso{
Other matrix functions: 
\code{\link{cor2cov}()},
\code{\link{low2sym}()},
\code{\link{up2sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{matrix functions}
\keyword{matrix}
