% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{up2sym}
\alias{up2sym}
\title{Upper Triangle to Symmetric}
\usage{
up2sym(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
Returns a symmetric matrix.
}
\description{
Creates a symmetric matrix from the upper triangle
of a square matrix.
}
\examples{
x <- matrix(NA, ncol = 4, nrow = 4)
x[upper.tri(x, diag = TRUE)] <- 1:10
up2sym(x = x)
}
\seealso{
Other matrix functions: 
\code{\link{cor2cov}()},
\code{\link{low2sym}()},
\code{\link{tr}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{matrix functions}
\keyword{matrix}
