% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorization.R
\name{vec}
\alias{vec}
\title{vec}
\usage{
vec(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\description{
Vectorization.
}
\details{
\deqn{
  \mathrm{vec} \left( \mathbf{A} \right)
  = \left[a_{1, 1}, \cdots, a_{m, 1}, a_{1, 2}, \cdots, a_{m, 2}, \cdots a_{1, n}, \cdots, a_{m, n} \right]^{T}
  }
}
\examples{
A <- matrix(
  data = c(1, 2, 2, 1),
  ncol = 2
)
vec(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)}{Wikipedia: Vectorization}
}
\author{
Ivan Jacob Agaloos Pesigan
}
