% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorization.R
\name{vech}
\alias{vech}
\title{vech}
\usage{
vech(X)
}
\arguments{
\item{X}{Square numeric matrix.}
}
\description{
Half-vectorization.
}
\details{
\deqn{
    \mathrm{vech} \left( \mathbf{A} \right)
    = \left[ a_{1, 1}, \cdots, a_{n, 1}, a_{2, 2}, \cdots, a_{n, 2}, \cdots,
    a_{n - 1, n - 1}, a_{n, n - 1}, a_{n, n} \right]^{T} .
  }
}
\examples{
A <- matrix(
  data = c(1, 2, 2, 1),
  ncol = 2
)
vec(A)
vech(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)#Half-vectorization}{Wikipedia: Half-vectorization}
}
\author{
Ivan Jacob Agaloos Pesigan
}
