% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{vech2tri}
\alias{vech2tri}
\title{Vector to Triangular Matrix}
\usage{
vech2tri(x, lower = TRUE, diag = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{lower}{Logical.
If \code{TRUE}, creates a lower triangular matrix.
If \code{FALSE}, creates an upper triangular matrix.}

\item{diag}{Logical.
If \code{TRUE},
values of the diagonal as supplied by \code{x}.
If \code{FALSE},
diagonals are set to zero and
\code{x} fills only the off-diagonal elements of the matrix
producing a strictly triangular matrix.}
}
\description{
Fills a matrix with a vector that represents the triangular elements.
The function can produce a strictly triangular matrix,
that is, the diagonal elements are set to zero
or a triangular matrix with diagonal values
supplied by the argument \code{x}.
Note that the matrix is filled by column.
}
\examples{
# Stricly lower triangular matrix
vech2tri(1:10, lower = TRUE, diag = FALSE)
# Lower triangular matrix with diagonal and off-diagonal values
# supplied by the argument x
vech2tri(1:10, lower = TRUE, diag = TRUE)
# Stricly upper triangular matrix
vech2tri(1:10, lower = FALSE, diag = FALSE)
# Upper triangular matrix with diagonal and off-diagonal values
# supplied by the argument x
vech2tri(1:10, lower = FALSE, diag = TRUE)
}
\author{
Ivan Jacob Agaloos Pesigan
}
