% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{is.positive.definite}
\alias{is.positive.definite}
\title{Positive Definite}
\usage{
is.positive.definite(x, tol = 1e-08)
}
\arguments{
\item{x}{A numeric square matrix.}

\item{tol}{Tolerance.}
}
\value{
Returns \code{FALSE}
if any of the eigenvalues are less than or equal to 0.
}
\description{
Checks if eigenvalues in a square matrix are positive.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is.positive.definite(Sigma)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{matrix}
