% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{low2sym}
\alias{low2sym}
\title{Lower Triangle to Symmetric}
\usage{
low2sym(X)
}
\arguments{
\item{X}{Square numeric matrix.}
}
\value{
Returns a symmetric matrix.
}
\description{
Creates a symmetric matrix from the lower triangle
of a square matrix.
}
\examples{
X <- matrix(NA, ncol = 4, nrow = 4)
X[lower.tri(X, diag = TRUE)] <- 1:10
low2sym(X = X)
}
\seealso{
Other matrix functions: 
\code{\link{cor2cov}()},
\code{\link{tr}()},
\code{\link{up2sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{matrix functions}
\keyword{matrix}
