% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{up2sym}
\alias{up2sym}
\title{Upper Triangle to Symmetric}
\usage{
up2sym(X)
}
\arguments{
\item{X}{Numeric matrix.
A \eqn{p \times p} matrix.}
}
\value{
Returns a symmetric matrix.
}
\description{
Creates a symmetric matrix from the upper triangle
of a square matrix.
}
\examples{
X <- matrix(NA, ncol = 4, nrow = 4)
X[upper.tri(X, diag = TRUE)] <- 1:10
up2sym(X = X)
}
\seealso{
Other matrix functions: 
\code{\link{cor2cov}()},
\code{\link{low2sym}()},
\code{\link{tr}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{matrix functions}
\keyword{matrix}
